// src/types/pokemonType.ts
export interface Pokemon {
    id: number;
    name: string;
    base_experience: number;
    height: number;
    is_default: boolean;
    order: number;
    weight: number;
    abilities: Ability[];
    forms: NamedAPIResource[];
    game_indices: VersionGameIndex[];
    held_items: PokemonHeldItem[];
    location_area_encounters: string;
    moves: PokemonMove[];
    sprites: PokemonSprites;
    species: NamedAPIResource;
    stats: PokemonStat[];
    types: PokemonType[];
}

interface Ability {
    is_hidden: boolean;
    slot: number;
    ability: NamedAPIResource;
}

interface NamedAPIResource {
    name: string;
    url: string;
}

interface VersionGameIndex {
    game_index: number;
    version: NamedAPIResource;
}

interface PokemonHeldItem {
    item: NamedAPIResource;
    version_details: PokemonHeldItemVersion[];
}

interface PokemonHeldItemVersion {
    version: NamedAPIResource;
    rarity: number;
}

interface PokemonMove {
    move: NamedAPIResource;
    version_group_details: PokemonMoveVersion[];
}

interface PokemonMoveVersion {
    move_learn_method: NamedAPIResource;
    version_group: NamedAPIResource;
    level_learned_at: number;
}

interface PokemonSprites {
    front_default: string;
    front_shiny: string;
    front_female: string | null;
    front_shiny_female: string | null;
    back_default: string;
    back_shiny: string;
    back_female: string | null;
    back_shiny_female: string | null;
}

interface PokemonStat {
    stat: NamedAPIResource;
    effort: number;
    base_stat: number;
}

interface PokemonType {
    slot: number;
    type: NamedAPIResource;
}

export type PokemonTypeColors = {
    bug: string;
    dark: string;
    dragon: string;
    electric: string;
    fairy: string;
    fighting: string;
    fire: string;
    flying: string;
    ghost: string;
    grass: string;
    ground: string;
    ice: string;
    normal: string;
    poison: string;
    psychic: string;
    rock: string;
    steel: string;
    water: string;
};